Reorganized aimbots
