+++ hooks/paint_traverse/paint_traverse.h
@@ -11,7 +11,7 @@ void draw_player_esp();
 void init_render_queue();
 void add_to_render_queue(const wchar_t *text, int x, int y, struct vec3_t color, float data);
 void add_bbox_decorator(const wchar_t *text, struct vec3_t color, void *entity);
-void add_timer(int x, int y, struct vec3_t color, float curtime, float starttime, float endtime);
+void add_timer(struct vec3_t world_pos, struct vec3_t color, float starttime, float endtime);
 void draw_render_queue();
 void draw_bbox_decorators(int start_x, int start_y, void *entity);
 void draw_timer_queue();
